@@ -21,7 +21,7 @@
 
 #include <vector>
 #include <string>
-#include <boost/optional.hpp>
+#include <optional>
 
 #include "util/blob.hpp"
 
@@ -33,7 +33,7 @@ private:
   std::string m_program;
   bool m_absolute_path;
   std::vector<std::string> m_arguments;
-  boost::optional<std::string> m_working_directory;
+  std::optional<std::string> m_working_directory;
 
   std::vector<char> m_stdout_vec;
   std::vector<char> m_stderr_vec;
